Move 'sendMessageToChildControls' from NppDarkMode.h to dpiManagerV2.h
