+++ PowerEditor/src/NppDarkMode.cpp
@@ -2981,77 +2981,6 @@ namespace NppDarkMode
 		SetWindowSubclass(hwnd, WindowNotifySubclass, g_windowNotifySubclassID, 0);
 	}
 
-	// currently send message only to selected buttons; listbox and edit controls with scrollbars
-	void sendMessageToChildControls(HWND hwndParent, UINT msg, WPARAM wParam, LPARAM lParam)
-	{
-		struct WMessage
-		{
-			UINT _msg = 0;
-			WPARAM _wParam = 0;
-			LPARAM _lParam = 0;
-		};
-
-		struct WMessage p { msg, wParam, lParam };
-
-		::EnumChildWindows(hwndParent, [](HWND hwnd, LPARAM childLParam) WINAPI_LAMBDA->BOOL{
-			auto & p = *reinterpret_cast<WMessage*>(childLParam);
-			constexpr size_t classNameLen = 32;
-			TCHAR className[classNameLen]{};
-			::GetClassName(hwnd, className, classNameLen);
-			auto style = ::GetWindowLongPtr(hwnd, GWL_STYLE);
-
-			if (wcscmp(className, WC_BUTTON) == 0)
-			{
-				switch (style & BS_TYPEMASK)
-				{
-					case BS_CHECKBOX:
-					case BS_AUTOCHECKBOX:
-					case BS_3STATE:
-					case BS_AUTO3STATE:
-					case BS_RADIOBUTTON:
-					case BS_AUTORADIOBUTTON:
-					{
-						if ((style & BS_PUSHLIKE) != BS_PUSHLIKE)
-						{
-							::SendMessage(hwnd, p._msg, p._wParam, p._lParam);
-						}
-						break;
-					}
-
-					default:
-					{
-						break;
-					}
-				}
-				return TRUE;
-			}
-
-			if (wcscmp(className, WC_EDIT) == 0)
-			{
-				bool hasScrollBar = ((style & WS_HSCROLL) == WS_HSCROLL) || ((style & WS_VSCROLL) == WS_VSCROLL);
-				if (hasScrollBar)
-				{
-					::SendMessage(hwnd, p._msg, p._wParam, p._lParam);
-				}
-				return TRUE;
-			}
-
-			if (wcscmp(className, WC_LISTBOX) == 0)
-			{
-				if ((style & LBS_COMBOBOX) != LBS_COMBOBOX)
-				{
-					bool hasScrollBar = ((style & WS_HSCROLL) == WS_HSCROLL) || ((style & WS_VSCROLL) == WS_VSCROLL);
-					if (hasScrollBar)
-					{
-						::SendMessage(hwnd, p._msg, p._wParam, p._lParam);
-					}
-				}
-				return TRUE;
-			}
-			return TRUE;
-			}, reinterpret_cast<LPARAM>(&p));
-	}
-
 	void setDarkTitleBar(HWND hwnd)
 	{
 		constexpr DWORD win10Build2004 = 19041;
